requirements.txt
# # FastAPI
# fastapi==0.111.0
# uvicorn[standard]==0.29.0
# httpx==0.27.0
# python-jose[cryptography]==3.3.0
# bcrypt==4.1.3
# python-multipart==0.0.9
# pydantic-settings==2.2.1

# # SQLAlchemy (async)
# sqlalchemy[asyncio]==2.0.30
# asyncpg==0.29.0
# alembic==1.13.1

# # Django
# django==5.0.4
# psycopg2-binary==2.9.9
# djangorestframework==3.15.1

# # Shared
# python-dotenv==1.0.1

# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic-settings==2.2.1

# SQLAlchemy (async)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Django
django==5.0.4
psycopg2-binary==2.9.9
djangorestframework==3.15.1
gunicorn==22.0.0
whitenoise==6.7.0

# Shared
python-dotenv==1.0.1